added the beginings of all the bio-stable genes, as well as updating the catagory
made the general genes for the bio-stabalized genes, along with improving the categories order.

The reason this is dupplicated is because when I first Commited it, I found out I didin't include the category, tried to un-commited, eventually wound up reverting files (but not commit). I just decided to leave it and re-commit this