new base
added a gene that will be used as a basis for all classes

the idea is that, instead of having a base that all genes uses, I have a base that is used by 2 bases. The first base is for RoM classes (so exclusions), the second is for traits (doesn't contain exclusions)

a future plan (improved exclusions 2) will have a base for mages and a base for fighters, but that's later.